MODIFY: 스토리피드 데이터 js파일로 저장하여 수정
